Create YPA2015 Manual Resize A.H.off
Testing A manual Resize Function Unfortunately It will constantly Reset and the resize icon appears and can still be interacted with in Fullscreen mode and will overlap with other page objects when resizing and resize gets locked when player height is changed in the style. not going to add feature till problems are fixed.